Add grammar for conditionals and return statements
The unfilled arrows mark control flow. The right facing arrow indicates the
start of control flow, (the `if` in an if statement). The box marks that the
conditional chain has ended (`else`). The diagonal arrows mean that it
continues (`elif`). The `return` arrow faces out because we go out of the
current scope. Break does the same thing, but only one level of scope, so it's
not solid.